move show test function to new file
